Making saving

+++ source/Area/sprites/sprite.h
@@ -3,7 +3,7 @@
 
 #include <stdlib.h>
 #include <SDL.h>
-
+#include <stdint.h>
 
 typedef struct SDL_Point Point;
 
@@ -56,7 +56,7 @@ typedef struct Sprite {
     SDL_Texture *spriteTexture;
     ParametresSprite arguments;
     SDL_RendererFlip reflaction;
-    unsigned long int id;
+    unsigned int id;
     unsigned int type;    
 } Sprite;
